misc/HtmlTestSuite/results/test-0212.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>CSS Container Queries Test: viewport units</title>
<link href="https://drafts.csswg.org/css-contain-3/#size-container" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="support/cq-testcommon.js"/>
<style>
  #vw { container-type: inline-size; width: 10vw; }
  #vh { container-type: inline-size; width: 10vh; }

  @container (min-width: 10vw) {
    #vw span { color: green }
  }
  @container (min-width: 11vw) {
    #vw span { color: red }
  }
  @container (min-width: 10vh) {
    #vh span { color: green }
  }
  @container (min-width: 11vh) {
    #vh span { color: red }
  }
</style>
</head><body><div id="vw"><span>Green</span></div>
<div id="vh"><span>Green</span></div>
<script>
  setup(() =&gt; assert_implements_container_queries());

  const green = "rgb(0, 128, 0)";

  test(() =&gt; assert_equals(getComputedStyle(vw.firstChild).color, green), "Match width with vw");
  test(() =&gt; assert_equals(getComputedStyle(vh.firstChild).color, green), "Match width with vh");
</script>
</body></html>
